- Stored most of the parameters as static variables in a new class called PropertySet - Created a schema and a sample xml file which can be used for initialising all the parameters of the model -Added a new Mason_15.jar file with more functionality. -Added more jar files for charts, media,etc, which can be used later as per requirements
NOTE : I had a few reference problems when making this commit. So please tell me if things mess up because of this. Thanks!

+++ src/agent/clustering/ClusteredSpace.java
@@ -31,17 +31,18 @@ import sim.util.Double2D;
 public final class ClusteredSpace extends RVOSpace {
 
     public static final double ALPHA = 3;
-    public static final int CLUSTER_DIAMETER = (int) Math.round(ALPHA * 2 * RVOAgent.RADIUS * 100.0);
+    public static final int CLUSTER_DIAMETER = (int) Math.round(
+                                ALPHA * 2 * RVOAgent.RADIUS * 100.0);
     public static final int NUMBER_OF_CLUSTERING_SPACES = 3;
     protected static int numberOfClusteringSpaces;
     protected ArrayList<ClusteredAgent>[] clusteredAgents;
     protected Continuous2D[] clusteringLayers;
 
-    public ClusteredSpace(int xSize, int ySize, float gridSize, RVOModel rm) {
+    public ClusteredSpace(int xSize, int ySize, double gridSize, RVOModel rm) {
         this(xSize, ySize, gridSize, rm, NUMBER_OF_CLUSTERING_SPACES);
     }
 
-    public ClusteredSpace(int xSize, int ySize, float gridSize, RVOModel rm,
+    public ClusteredSpace(int xSize, int ySize, double gridSize, RVOModel rm,
             int numberOfLayers) {
         super(xSize, ySize, gridSize, rm);
         numberOfClusteringSpaces = numberOfLayers;
